ENH: Add support for BZIP2 commpression from NrrdIO
bzip2 is nearly as common as zlib on build environments.
This initial patch set universally adds bz2 dependancies
to the NrrdIO library.

A followup set of changes may be necessary to make bzip2
a conditionally supported feature, but this is the first step
that will allow other testing to occur for integration with ITK.