STYLE: Minimize code changes between teem and NrrdIO
The changed files are stubs that remove most of the
functionality for file formats that are not part of
the NrrdIO libary.

The API should be consistent in both teem and NrrdIO,
but have diverged in some cases.

Make the argument variable names consistent (in some cases
and _ prefix was missing relative to the upstream teem).

Make functions static. Upstream teem identified many functions
that do not need to be part of the public API.  This will help
with minimizing the name mangling that is exposed.

Fix some static const char [] for function names to match the
function names used upstream.